Improve error handling

See merge request Concordium/p2p-client!999

+++ concordium-node/src/p2p/bans.rs
@@ -57,15 +57,16 @@ impl P2PNode {
     pub fn ban_node(&self, peer: BanId) -> Fallible<()> {
         info!("Banning node {:?}", peer);
 
-        let mut store_key = Vec::new();
-        peer.serial(&mut store_key);
-        {
-            let ban_kvs_env = safe_read!(self.kvs)?;
+        if let Ok(ban_kvs_env) = self.kvs.read() {
+            let mut store_key = Vec::new();
+            peer.serial(&mut store_key);
             let ban_store = ban_kvs_env.open_single(BAN_STORE_NAME, StoreOptions::create())?;
             let mut writer = ban_kvs_env.write()?;
             // TODO: insert ban expiry timestamp as the Value
             ban_store.put(&mut writer, store_key, &Value::U64(0))?;
             writer.commit()?;
+        } else {
+            bail!("Couldn't ban a peer: couldn't obtain a lock over the kvs");
         }
 
         match peer {
@@ -93,15 +94,16 @@ impl P2PNode {
     pub fn unban_node(&self, peer: BanId) -> Fallible<()> {
         info!("Unbanning node {:?}", peer);
 
-        let mut store_key = Vec::new();
-        peer.serial(&mut store_key);
-        {
-            let ban_kvs_env = safe_read!(self.kvs)?;
+        if let Ok(ban_kvs_env) = self.kvs.read() {
+            let mut store_key = Vec::new();
+            peer.serial(&mut store_key);
             let ban_store = ban_kvs_env.open_single(BAN_STORE_NAME, StoreOptions::create())?;
             let mut writer = ban_kvs_env.write()?;
             // TODO: insert ban expiry timestamp as the Value
             ban_store.delete(&mut writer, store_key)?;
             writer.commit()?;
+        } else {
+            bail!("Couldn't unban a peer: couldn't obtain a lock over the kvs");
         }
 
         if !self.config.no_trust_bans {
@@ -113,40 +115,48 @@ impl P2PNode {
 
     /// Checks whether a specified id has been banned.
     pub fn is_banned(&self, peer: BanId) -> Fallible<bool> {
-        let ban_kvs_env = safe_read!(self.kvs)?;
-        let ban_store = ban_kvs_env.open_single(BAN_STORE_NAME, StoreOptions::create())?;
+        if let Ok(ban_kvs_env) = self.kvs.read() {
+            let ban_store = ban_kvs_env.open_single(BAN_STORE_NAME, StoreOptions::create())?;
+            let ban_reader = ban_kvs_env.read()?;
+            let mut store_key = Vec::new();
+            peer.serial(&mut store_key);
 
-        let ban_reader = ban_kvs_env.read()?;
-        let mut store_key = Vec::new();
-        peer.serial(&mut store_key);
-
-        Ok(ban_store.get(&ban_reader, store_key)?.is_some())
+            Ok(ban_store.get(&ban_reader, store_key)?.is_some())
+        } else {
+            bail!("Couldn't check if a peer is banned: couldn't obtain a lock over the kvs");
+        }
     }
 
     /// Obtains the list of banned nodes.
     pub fn get_banlist(&self) -> Fallible<Vec<BanId>> {
-        let ban_kvs_env = safe_read!(self.kvs)?;
-        let ban_store = ban_kvs_env.open_single(BAN_STORE_NAME, StoreOptions::create())?;
+        if let Ok(ban_kvs_env) = self.kvs.read() {
+            let ban_store = ban_kvs_env.open_single(BAN_STORE_NAME, StoreOptions::create())?;
 
-        let ban_reader = ban_kvs_env.read()?;
-        let ban_iter = ban_store.iter_start(&ban_reader)?;
+            let ban_reader = ban_kvs_env.read()?;
+            let ban_iter = ban_store.iter_start(&ban_reader)?;
 
-        let mut banlist = Vec::new();
-        for entry in ban_iter {
-            let (mut id_bytes, _expiry) = entry?;
-            let node_to_ban = BanId::deserial(&mut id_bytes)?;
-            banlist.push(node_to_ban);
+            let mut banlist = Vec::new();
+            for entry in ban_iter {
+                let (mut id_bytes, _expiry) = entry?;
+                let node_to_ban = BanId::deserial(&mut id_bytes)?;
+                banlist.push(node_to_ban);
+            }
+
+            Ok(banlist)
+        } else {
+            bail!("Couldn't get the banlist: couldn't obtain a lock over the kvs");
         }
-
-        Ok(banlist)
     }
 
     /// Lifts all existing bans.
     pub fn clear_bans(&self) -> Fallible<()> {
-        let kvs_env = safe_read!(self.kvs)?;
-        let ban_store = kvs_env.open_single(BAN_STORE_NAME, StoreOptions::create())?;
-        let mut writer = kvs_env.write()?;
-        ban_store.clear(&mut writer)?;
-        into_err!(writer.commit())
+        if let Ok(kvs_env) = self.kvs.read() {
+            let ban_store = kvs_env.open_single(BAN_STORE_NAME, StoreOptions::create())?;
+            let mut writer = kvs_env.write()?;
+            ban_store.clear(&mut writer)?;
+            writer.commit().map_err(|e| e.into())
+        } else {
+            bail!("Couldn't clear the bans: couldn't obtain a lock over the kvs");
+        }
     }
 }
